+++ reghelp_client/__init__.py
@@ -41,7 +41,7 @@ from .models import (
     VoipStatusResponse,
 )
 
-__version__ = "1.6.1"
+__version__ = "1.7.0"
 __all__ = [
     "RegHelpClient",
     "BalanceResponse",